First approach for extracting methods for lib/pghero module
